requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-groq
langgraph
ddgs
wikipedia
arxiv
